Task 4: Move FileSize/ByteSize to int64 in v4 and core
os.FileInfo.Size() is int64, but the on-disk ByteSize field was uint64, so
every bridge between them was a gosec-G115-flagged signed<->unsigned
conversion. Task 3.3 annotated 7 of these on the FileSize/.Size() path with
//nolint:gosec. This task reinterprets the on-disk size field as signed
int64 throughout pkg/format and core, retiring those 7 suppressions by
removing the conversions rather than re-suppressing.

It is a signedness reinterpretation of an already-8-byte field, NOT a format
bump: CurrentIndexVersion stays 4 and v4 indices round-trip byte-identical.

Changes:
- pkg/format/vocabulary.go: ByteSize alias uint64->int64 (single load-bearing
  edit). Entry/v2/v3/v4 struct fields, codec Get/SetFileSize, and transcode.go
  follow transparently via the alias.
- Interfaces + 7 implementors retyped to int64 (BinaryEntryInterface.FileSize,
  FilterEntry.FileSize and their adapters/scan/skiplist/mmap/mock impls).
  needsHash compares FileSize inline as int64 (pulled out of the uint64-typed
  field slice) so no int64->uint64 conversion is reintroduced.
- Signed size-threshold plumbing: ParseSizeBound->int64 (with a leading-sign
  guard preserving the unsigned-magnitude grammar ParseUint enforced),
  SizeBoundString, FilterOptions.MinSize/MaxSize, MinSizeTest.Min,
  MaxSizeTest.Max, FindOptions.ExactSize, dedupeDefaultMinSize. EntryInfo/
  EntryJSON FileSize, ValidationConfig.MaxFileSize, and the dcfhfix repair
  parse (parseInt64) all signed end-to-end.
- Corruption floor: a negative FileSize (the only value the signed reinterpret
  can newly produce, from a corrupt/>=2^63 legacy field) is rejected
  fail-closed by validateFileSizeBounds (recovery.go) and the dcfhfind
  corruption checks, mirroring the pre-1885 wall-time underflow guard.
- 7 size G115 suppressions retired by deleting the cast. golangci-lint run
  ./... reports 0 G115 whole-tree; the --new staged gate is clean.

All four success criteria verified end-to-end (g-testing-exec.md): SC1 v4
byte-identical + large-positive round-trip, SC2 0 G115 whole-tree + --new
clean + the 7 lines suppression-free, SC3 legacy v2/v3 decode unchanged +
negative-size fail-closed, SC4 full suite + race green. 0 escaped defects.

The fsdedupe byte totals (distinct uint64 type with its own JSON contract)
and the dupes.go formatting casts are deliberately out of scope; their 3
suppressions are retained, recorded not deferred.

Closes the int64 backlog item ("Move FileSize/ByteSize to int64 in v4 +
core"; carried the stale "subtask 3.4" naming, landed as top-level Task 4).

+++ pkg/binary_entry_index_file_mmap_test.go
@@ -48,7 +48,7 @@ func createBEIndexFileMmap(t *testing.T, testData *TestEntryData) BinaryEntryInt
 
 	mockInfo := &mockFileInfo{
 		name:    "test_file.txt",
-		size:    int64(testData.FileSize),
+		size:    testData.FileSize,
 		mode:    os.FileMode(testData.Mode),
 		modTime: time.Unix(1234567900, 0),
 	}
